+++ client/public/components/index.js
@@ -2,4 +2,5 @@ import { Header } from "./header";
 import { Nav } from "./nav";
 import { Footer } from "./footer";
 import { Card } from "./card";
-export { Header, Nav, Footer, Card };
+import { ErrorPage } from './error-page';
+export { Header, Nav, Footer, Card, ErrorPage };
